feat: 배송 상세 조회 응용, 인프라계층 개발

+++ delivery-service/src/main/java/com/firstlogistics/deliverservice/application/dto/command/CreateDeliveryCommand.java
@@ -34,7 +34,7 @@ public record CreateDeliveryCommand(
 		UUID productId,
 		String productName,
 		int quantity,
-		int price
+		Long price
 	) {}
 
 	public static CreateDeliveryCommand from(OrderAcceptedEvent event) {
